Desktop/WORK/professional_investigators_cybertrace/web_requirements.txt
# Professional Phone Intelligence Web Application - Requirements
# Version: 1.0.0

# Core Web Framework
Flask>=2.3.0
Flask-SocketIO>=5.3.0
Flask-CORS>=4.0.0

# WebSocket Support
python-socketio>=5.8.0
eventlet>=0.33.0

# File Handling
Werkzeug>=2.3.0

# Async Support
asyncio-mqtt>=0.16.0
aiohttp>=3.8.0

# Existing Phone Intelligence Dependencies
pyyaml>=6.0
requests>=2.28.0
dnspython>=2.3.0
python-whois>=0.8.0
cryptography>=40.0.0
bcrypt>=4.0.0
phonenumbers>=8.13.0

# Optional Enhanced Features
pillow>=9.0.0
python-magic>=0.4.27
xattr>=0.10.1

# Development and Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

# Production Server (Optional)
gunicorn>=20.1.0
gevent>=22.10.0

# Monitoring and Logging
prometheus-client>=0.16.0